Feature/update custom master 10 2021 (#6)
* remove always true condition in if statement

* Improve speed of drw_text when provided with large strings

Calculates len & ew in drw_font_getexts loop by incrementing instead of
decrementing; as such avoids proportional increase in time spent in loop
based on provided strings size.

* add support for more keypad keys

The keypad Enter key was already supported. On some keyboard layouts like my
laptop the page-up and page-down key is more comfortable to use.
This adds a few lines but no complexity.

* Revert "Improve speed of drw_text when provided with large strings"

This reverts commit c585e8e498ec6f9c423ab8ea07cf853ee5b05fbe.

It causes issues with truncation of characters when the text does not fit and
so on.  The patch should be reworked and properly tested.

Co-authored-by: Guilherme Janczak <guilherme.janczak@yandex.com>
Co-authored-by: Miles Alan <m@milesalan.com>
Co-authored-by: Hiltjo Posthuma <hiltjo@codemadness.org>
Co-authored-by: Paul van de Ven <paul@gutstickets.org>